# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PaymentInfo {
  upiId: string;
  qrImageUrl: string;
  minDeposit: number;
}

export interface BalanceInfo {
  telegramUserId: number;
  balance: number;
}

export interface UtrVerifyInput {
  telegramUserId: number;
  /** @minLength 6 */
  utr: string;
}

export interface UtrVerifyResult {
  success: boolean;
  amount: number;
  oldBalance: number;
  newBalance: number;
  utr: string;
  ref: string;
  date: string;
}

export interface UpiError {
  error: string;
}

export interface AdminCheckResult {
  isAdmin: boolean;
}

export interface UpdatePaymentInfoInput {
  /** @minLength 1 */
  upiId: string;
  /** @minLength 1 */
  qrImageUrl: string;
  /** @minimum 0 */
  minDeposit: number;
}

export interface VerifyDeviceInput {
  /** @minLength 8 */
  fingerprint: string;
}

export type VerifyDeviceResultReason = typeof VerifyDeviceResultReason[keyof typeof VerifyDeviceResultReason];


export const VerifyDeviceResultReason = {
  new_device: 'new_device',
  already_verified: 'already_verified',
  duplicate_device: 'duplicate_device',
} as const;

export interface VerifyDeviceResult {
  verified: boolean;
  reason: VerifyDeviceResultReason;
}

export interface CreateDepositSessionInput {
  /** @minimum 1 */
  amountInr: number;
}

export type DepositSessionStatusProperty = typeof DepositSessionStatusProperty[keyof typeof DepositSessionStatusProperty];


export const DepositSessionStatusProperty = {
  pending: 'pending',
} as const;

export interface DepositSession {
  orderId: string;
  amountInr: number;
  amountUsd: number;
  upiId: string;
  upiLink: string;
  qrImageUrl: string;
  status: DepositSessionStatusProperty;
  expiresAt: string;
}

export type DepositSessionStatusStatus = typeof DepositSessionStatusStatus[keyof typeof DepositSessionStatusStatus];


export const DepositSessionStatusStatus = {
  pending: 'pending',
  paid: 'paid',
  expired: 'expired',
  failed: 'failed',
} as const;

export interface DepositSessionStatus {
  orderId: string;
  status: DepositSessionStatusStatus;
  amountInr: number;
  amountUsd: number;
  newBalance?: number;
  utr?: string;
  ref?: string;
  date?: string;
  error?: string;
}