# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: upi
    description: UPI deposit verification for the Telegram mini app
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /upi/payment-info:
    get:
      operationId: getPaymentInfo
      tags: [upi]
      summary: Get the admin's UPI QR code and UPI ID
      description: Returns the payment details the mini app should display for deposits
      responses:
        "200":
          description: Payment info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentInfo"
  /upi/balance/{telegramUserId}:
    get:
      operationId: getUpiBalance
      tags: [upi]
      summary: Get a user's bot balance
      parameters:
        - name: telegramUserId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalanceInfo"
        "400":
          description: Invalid user id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpiError"
  /upi/admin/check:
    get:
      operationId: checkAdmin
      tags: [upi]
      summary: Check whether the calling Telegram user is the configured admin
      description: >-
        Verifies the caller's Telegram identity from the signed initData header
        and reports whether they match the bot's configured admin id.
      responses:
        "200":
          description: Admin check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCheckResult"
  /upi/admin/payment-info:
    put:
      operationId: updatePaymentInfo
      tags: [upi]
      summary: Update the UPI QR code, UPI ID, and minimum deposit (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePaymentInfoInput"
      responses:
        "200":
          description: Updated payment info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentInfo"
        "401":
          description: Missing or invalid Telegram authentication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpiError"
        "403":
          description: Caller is not the configured admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpiError"
  /upi/verify-device:
    post:
      operationId: verifyDevice
      tags: [upi]
      summary: Register/check a device fingerprint to prevent multi-account referral abuse
      description: >-
        Verifies the caller's Telegram identity from the signed initData header,
        then checks whether the given device fingerprint is already bound to a
        different Telegram account. If the fingerprint is new (or already bound
        to this same account), it is (re)bound to this account and verification
        succeeds. If it belongs to a different account, verification fails so
        the bot can refuse to pay out a referral bonus for this device.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyDeviceInput"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyDeviceResult"
        "401":
          description: Missing or invalid Telegram authentication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpiError"
  /upi/verify:
    post:
      operationId: verifyUtr
      tags: [upi]
      summary: Verify a UTR/transaction id and credit the user's balance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UtrVerifyInput"
      responses:
        "200":
          description: Payment verified and balance credited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UtrVerifyResult"
        "400":
          description: Invalid input, already-used UTR, or payment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpiError"
  /upi/deposit-session:
    post:
      operationId: createDepositSession
      tags: [upi]
      summary: Create an amount-specific UPI deposit session
      description: Creates a unique order and UPI deep link that the mini app can monitor for automatic payment verification.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepositSessionInput"
      responses:
        "201":
          description: Deposit session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositSession"
        "400":
          description: UPI automatic verification is not configured or amount is invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpiError"
        "401":
          description: Missing or invalid Telegram authentication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpiError"
  /upi/deposit-session/{orderId}:
    get:
      operationId: getDepositSession
      tags: [upi]
      summary: Get the live status of a UPI deposit session
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deposit session status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositSessionStatus"
        "401":
          description: Missing or invalid Telegram authentication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpiError"
        "403":
          description: Session belongs to another Telegram user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpiError"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpiError"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    PaymentInfo:
      type: object
      properties:
        upiId:
          type: string
        qrImageUrl:
          type: string
        minDeposit:
          type: number
      required:
        - upiId
        - qrImageUrl
        - minDeposit
    BalanceInfo:
      type: object
      properties:
        telegramUserId:
          type: number
        balance:
          type: number
      required:
        - telegramUserId
        - balance
    UtrVerifyInput:
      type: object
      properties:
        telegramUserId:
          type: number
        utr:
          type: string
          minLength: 6
      required:
        - telegramUserId
        - utr
    UtrVerifyResult:
      type: object
      properties:
        success:
          type: boolean
        amount:
          type: number
        oldBalance:
          type: number
        newBalance:
          type: number
        utr:
          type: string
        ref:
          type: string
        date:
          type: string
      required:
        - success
        - amount
        - oldBalance
        - newBalance
        - utr
        - ref
        - date
    UpiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    AdminCheckResult:
      type: object
      properties:
        isAdmin:
          type: boolean
      required:
        - isAdmin
    UpdatePaymentInfoInput:
      type: object
      properties:
        upiId:
          type: string
          minLength: 1
        qrImageUrl:
          type: string
          minLength: 1
        minDeposit:
          type: number
          minimum: 0
      required:
        - upiId
        - qrImageUrl
        - minDeposit
    VerifyDeviceInput:
      type: object
      properties:
        fingerprint:
          type: string
          minLength: 8
      required:
        - fingerprint
    VerifyDeviceResult:
      type: object
      properties:
        verified:
          type: boolean
        reason:
          type: string
          enum: [new_device, already_verified, duplicate_device]
      required:
        - verified
        - reason
    CreateDepositSessionInput:
      type: object
      properties:
        amountInr:
          type: number
          minimum: 1
      required:
        - amountInr
    DepositSession:
      type: object
      properties:
        orderId:
          type: string
        amountInr:
          type: number
        amountUsd:
          type: number
        upiId:
          type: string
        upiLink:
          type: string
        qrImageUrl:
          type: string
        status:
          type: string
          enum: [pending]
        expiresAt:
          type: string
      required:
        - orderId
        - amountInr
        - amountUsd
        - upiId
        - upiLink
        - qrImageUrl
        - status
        - expiresAt
    DepositSessionStatus:
      type: object
      properties:
        orderId:
          type: string
        status:
          type: string
          enum: [pending, paid, expired, failed]
        amountInr:
          type: number
        amountUsd:
          type: number
        newBalance:
          type: number
        utr:
          type: string
        ref:
          type: string
        date:
          type: string
        error:
          type: string
      required:
        - orderId
        - status
        - amountInr
        - amountUsd