# lib/api-zod/src/generated/types/createDepositSessionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateDepositSessionInput {
  /** @minimum 1 */
  amountInr: number;
}