# lib/api-zod/src/generated/types/depositSession.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DepositSessionStatusProperty } from './depositSessionStatusProperty';

export interface DepositSession {
  orderId: string;
  amountInr: number;
  amountUsd: number;
  upiId: string;
  upiLink: string;
  qrImageUrl: string;
  status: DepositSessionStatusProperty;
  expiresAt: string;
}