# lib/api-zod/src/generated/types/depositSessionStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DepositSessionStatusStatus } from './depositSessionStatusStatus';

export interface DepositSessionStatus {
  orderId: string;
  status: DepositSessionStatusStatus;
  amountInr: number;
  amountUsd: number;
  newBalance?: number;
  utr?: string;
  ref?: string;
  date?: string;
  error?: string;
}